pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nvbroadcast"
version = "1.0.2"
description = "NV Broadcast - Unofficial NVIDIA Broadcast for Linux and other OS by doczeus - AI-powered virtual camera, background removal, and noise cancellation"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.11"
authors = [
    {name = "doczeus", email = "harshit@kshoonya.com"},
]
maintainers = [
    {name = "doczeus", email = "harshit@kshoonya.com"},
]
keywords = ["nvidia", "broadcast", "virtual-camera", "background-removal", "noise-cancellation", "linux", "doczeus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "numpy>=1.26",
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "mediapipe>=0.10",
    "onnxruntime-gpu>=1.16",
    "pyrnnoise>=0.4",
    "onnx>=1.14",
    "cupy-cuda12x",
    "psutil>=5.9",
    "nvidia-cublas-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-curand-cu12",
    "nvidia-cufft-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "openai-whisper>=20231117",
    "scipy>=1.11",
]

[project.urls]
Homepage = "https://github.com/Hkshoonya/nvidia-broadcast-linux"
Repository = "https://github.com/Hkshoonya/nvidia-broadcast-linux"
Issues = "https://github.com/Hkshoonya/nvidia-broadcast-linux/issues"

[project.optional-dependencies]
dev = ["pytest>=8.0"]
cuda = ["pycuda>=2024.1"]
maxine = ["nvidia-vfx"]
tensorrt = ["tensorrt", "tensorrt-cu12", "onnx"]

[project.scripts]
nvbroadcast = "nvbroadcast.__main__:main"
nvbroadcast-vcam = "nvbroadcast.vcam_service:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nvbroadcast.ui" = ["style.css"]

[tool.setuptools.data-files]
"share/applications" = ["data/com.doczeus.NVBroadcast.desktop"]
"share/metainfo" = ["data/com.doczeus.NVBroadcast.metainfo.xml"]
"share/icons/hicolor/scalable/apps" = ["data/icons/com.doczeus.NVBroadcast.svg"]
"share/nvbroadcast/backgrounds" = [
    "data/backgrounds/gradient_bg.png",
    "data/backgrounds/test_bg.png",
]
